refactor: update question scores
This commit introduces a new formula to compute the scores for multiple
questions. Previously the scores were hard coded to work with 4 options,
requiring manual adjustments for questions with a different number of
options. Now, the score will automatically adapt to the total number of
options and the number of correct options.